rename packages, using the std java naming convention
