+++ src/main/java/com/akkarimzai/exceptions/DatabindException.java
@@ -1,4 +1,4 @@
-package com.akkarimzai.Exceptions;
+package com.akkarimzai.exceptions;
 
 public class DatabindException extends Exception {
     public DatabindException(String message) {
