+++ src/main/java/com/akkarimzai/profiles/Serializers/MapPointSerializer.java
@@ -1,9 +1,9 @@
-package com.akkarimzai.Profiles.Serializers;
+package com.akkarimzai.profiles.Serializers;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.SerializerProvider;
 import com.fasterxml.jackson.databind.ser.std.StdSerializer;
-import com.akkarimzai.Entities.MapPoint;
+import com.akkarimzai.entities.MapPoint;
 
 import java.io.IOException;
