+++ src/main/java/com/akkarimzai/repositories/MapPointRepositoryFS.java
@@ -1,9 +1,9 @@
-package com.akkarimzai.Repositories;
+package com.akkarimzai.repositories;
 
-import com.akkarimzai.Exceptions.DatabindException;
-import com.akkarimzai.Exceptions.NotFoundException;
+import com.akkarimzai.exceptions.DatabindException;
+import com.akkarimzai.exceptions.NotFoundException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.akkarimzai.Entities.MapPoint;
+import com.akkarimzai.entities.MapPoint;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
