+++ src/test/java/com/akkarimzai/profiles/MapPointJsonMapperTest.java
@@ -1,6 +1,6 @@
-package com.akkarimzai.Profiles;
+package com.akkarimzai.profiles;
 
-import com.akkarimzai.Entities.MapPoint;
+import com.akkarimzai.entities.MapPoint;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
